Move enhanced check tests into controller test
